Revert "Revert "using maps, ludwig wanted this.""
This reverts commit e8543d5281be2238042ebeb3df3f976a8cc0fcd3.

+++ src/collatz_steps.cpp
@@ -1,13 +1,12 @@
 #include "collatz.h"
-#include <stdint.h>
 
 #ifdef EMSCRIPTEN_KEEPALIVE
 EMSCRIPTEN_KEEPALIVE
 #endif
-double collatz_steps(double i) {
-        double n;
+int collatz_steps(uint_least64_t i) {
+        int n;
         for(n = 0; i > 1; n++) {
-                if(((uint_least64_t)i) % 2 == 0) i /= 2;
+                if(i % 2 == 0) i /= 2;
                 else           i *= 3, i += 1;
         }
         return n;
